requirements.txt
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
loguru==0.5.3
mysqlclient @ file:///E:/PYTHON/Install/mysqlclient-1.4.2-cp37-cp37m-win32.whl
Pillow==8.1.0
python-dateutil==2.8.1
xlwings==0.20.8
